+++ src/But/Log/Field/FileName.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <string>
 #include <boost/filesystem/path.hpp>
 
 namespace But